no more warings

+++ string.c
@@ -109,7 +109,7 @@ size_t strspn(const char *str1, const char *str2) {
             break;
         }
     }
-    return s1 - str1;
+    return (size_t)(s1 - str1);
 }
 
 // Returns the length of the initial segment of str1 which consists entirely of characters not in str2
@@ -125,7 +125,7 @@ size_t strcspn(const char *str1, const char *str2) {
             break;
         }
     }
-    return s1 - str1;
+    return (size_t)(s1 - str1);
 }
 
 #endif
